+++ bot/cogs/account.py
@@ -4,10 +4,10 @@ import disnake
 from disnake.ext.commands import Cog, slash_command
 
 from bot.data import db_pb
-from bot.misc.constants import loginCommand
+from bot.misc.constants import loginCommand, switchCommand
 from bot.misc.penguin import Penguin
-from bot.data.pufflebot.users import Users, PenguinIntegrations
-from bot.handlers.buttons import Logout, Login
+from bot.data.pufflebot.user import User, PenguinIntegrations
+from bot.handlers.buttons import Buttons
 from bot.handlers.select import SelectPenguins
 from bot.misc.utils import getPenguinFromInter, getPenguinOrNoneFromUserId
 
@@ -25,7 +25,7 @@ class AccountManagementCommands(Cog):
     @slash_command(name="logout", description="Отвязать свой Discord аккаунт от своего пингвина")
     async def logout(self, inter: ApplicationCommandInteraction):
         p: Penguin = await getPenguinFromInter(inter)
-        user: Users = await Users.get(inter.user.id)
+        user: User = await User.get(inter.user.id)
         penguin_ids = await db_pb.select([PenguinIntegrations.penguin_id]).where(
             (PenguinIntegrations.discord_id == inter.user.id)).gino.all()
 
@@ -35,7 +35,7 @@ class AccountManagementCommands(Cog):
     @slash_command(name="switch", description="Сменить текущий аккаунт")
     async def switch(self, inter: ApplicationCommandInteraction):
         p: Penguin = await getPenguinOrNoneFromUserId(inter.user.id)
-        user: Users = await Users.get(inter.user.id)
+        user: User = await User.get(inter.user.id)
         penguin_ids = await db_pb.select([PenguinIntegrations.penguin_id]).where(
             (PenguinIntegrations.discord_id == inter.user.id)).gino.all()
 
@@ -62,6 +62,50 @@ class AccountManagementCommands(Cog):
         return await inter.send(
             f"Ваш текущий аккаунт: `{p.safe_name()}`. Какой аккаунт вы хотите сделать текущим?",
             view=view, ephemeral=True)
+
+
+class Login(disnake.ui.View):
+    def __init__(self):
+        super().__init__()
+
+    @disnake.ui.button(label="Войти", url="https://cpps.app/discord/login")
+    async def loginButton(self, button: disnake.ui.Button, inter: disnake.CommandInteraction):
+        ...
+
+
+class Logout(Buttons):
+    def __init__(self, inter: disnake.CommandInteraction, p, user, penguin_ids):
+        super().__init__(inter, timeout=None)
+        self.p = p
+        self.user = user
+        self.penguin_ids = penguin_ids
+
+    async def on_timeout(self):
+        ...
+
+    @disnake.ui.button(label="Отмена", style=disnake.ButtonStyle.gray, custom_id="cancel")
+    async def cancelButton(self, _, inter: disnake.CommandInteraction):
+        await self.disableAllItems()
+        await inter.send(f"Отменено", ephemeral=True)
+
+    @disnake.ui.button(label="Выйти", style=disnake.ButtonStyle.red, custom_id="logout")
+    async def logoutButton(self, _, inter: disnake.CommandInteraction):
+        await self.disableAllItems()
+        await PenguinIntegrations.delete.where(PenguinIntegrations.penguin_id == self.p.id).gino.status()
+        if len(self.penguin_ids) == 1:
+            await self.user.update(penguin_id=None).apply()
+            return await inter.send(f"Ваш аккаунт `{self.p.safe_name()}` успешно отвязан.", ephemeral=True)
+
+        if len(self.penguin_ids) == 2:
+            newCurrentPenguin: Penguin = await Penguin.get(self.penguin_ids[0][0])
+            await self.user.update(penguin_id=self.penguin_ids[0][0]).apply()
+            return await inter.send(
+                f"Ваш аккаунт `{self.p.safe_name()}` успешно отвязан. "
+                f"Сейчас ваш текущий аккаунт `{newCurrentPenguin.safe_name()}`", ephemeral=True)
+
+        return await inter.send(
+            f"Ваш аккаунт `{self.p.safe_name()}` успешно отвязан. "
+            f"Чтобы выбрать текущий аккаунт воспользуйтесь командой {switchCommand}", ephemeral=True)
 
 
 def setup(bot):
